Update all address schemas
