+++ src/schemas/address-schemas.ts
@@ -72,7 +72,7 @@ export const PersonAddressSchema: AddressSchemaIF = {
     maxLength: maxLength(50)
   },
   streetAddressAdditional: {
-    maxLength: maxLength(50)
+    maxLength: maxLength(105)
   },
   addressCity: {
     required,
